requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
ConfigArgParse==1.7.1
cryptography==45.0.3
defusedxml==0.8.0rc2
distro==1.9.0
Django==4.2.21
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-guardian==3.0.0
django-redis==6.0.0
django-silk==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
docker==7.1.0
drf-nested-routers==0.94.2
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
gprof2dot==2025.4.14
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.63
langchain-openai==0.3.18
langsmith==0.3.43
locust==2.34.0
MarkupSafe==3.0.2
msgpack==1.1.1
mysqlclient==2.2.7
oauthlib==3.2.2
openai==1.82.1
orjson==3.10.18
packaging==24.2
pluggy==1.6.0
psutil==7.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
pytest==8.3.5
python-dotenv==1.1.0
python3-openid==3.2.0
PyYAML==6.0.2
pyzmq==27.0.0
redis==6.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
setuptools==80.9.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0rc1
urllib3==2.4.0
Werkzeug==3.1.3
whitenoise==6.9.0
zipp==3.23.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
